requirements.txt
streamlit
google-generativeai
pymupdf
transformers
torch
accelerate # Often helpful with Hugging Face models
pandas
bitsandbytes